requirements.txt
discord
googletrans==4.0.0-rc1
python-dotenv